Move some type definitions to Flow
